requirements.txt
streamlit
opencv-contrib-python
torch
facenet-pytorch
pandas
numpy
scikit-learn
Pillow
